Z703 OS under construction
I've completed the prototype hardware.  There are a few issues, but I've got some stuff working:
2 digit display is working properly.
Channels 1 & 2 of attenuation are working.
SPI comms to VCA/Mix board is working.

Not working;
input from POTs/Faders are interacting with teach other ,  This may well be a configuration issue.
Channels 3 & 4 of the VCA/Mix - not yet working